requirements.txt
# GrantOps Dependencies

# LLM abstraction layer - supports OpenAI, Anthropic, and 100+ providers
litellm>=1.40.0

# YAML configuration parsing
pyyaml>=6.0

# HTTP client for API calls
httpx>=0.27.0

# Environment variable management
python-dotenv>=1.0.0
